normed main.c
