+++ src/execution/run_command.c
@@ -6,7 +6,7 @@
 /*   By: hmensah- <hmensah-@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/05/15 14:30:01 by hmensah-          #+#    #+#             */
-/*   Updated: 2025/05/15 19:20:56 by hmensah-         ###   ########.fr       */
+/*   Updated: 2025/05/16 16:47:27 by hmensah-         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -60,11 +60,13 @@ void	walk_ast(t_ast *ast, t_mshell *shell, t_allocs *allocs, t_table *table)
 		handle_others(ast, shell, allocs, table);
 }
 
-int	run_command(t_mshell *shell, t_allocs *allocs, t_table *table)
+int	run_command(t_mshell *shell, t_allocs *allocs, t_table *table,
+		t_result res)
 {
 	t_result	result;
 	char		*temp;
 
+	shell->ast = res.data.value;
 	result = get_env(table, "PATH");
 	if (result.is_error)
 		return (1);
